集成友盟、bugly

+++ PROJECTS/Models/WxGroupModel.h
@@ -8,11 +8,11 @@
 
 @interface WxGroupModel : BaseModel
 
-@property (assign, nonatomic) NSInteger gid;
+@property (assign, nonatomic) long gid;
 @property (strong, nonatomic) NSString *name;
 @property (strong, nonatomic) NSString *image;
-@property (assign, nonatomic) NSInteger currentCount;
-@property (assign, nonatomic) NSInteger maxCount;
+@property (assign, nonatomic) long currentCount;
+@property (assign, nonatomic) long maxCount;
 @property (assign, nonatomic) BOOL isManager;
 @property (strong, nonatomic) NSString *desc;
